Implements (partially) deserialization of a build step

+++ src/tty_util.rs
@@ -10,8 +10,10 @@ use nix::unistd::{isatty, dup};
 
 
 mod ffi {
-    pub static TIOCSPGRP: u64 = 0x5410;
-    pub static TIOCGPGRP: u64 = 0x540F;
+    use libc::{c_int};
+
+    pub static TIOCSPGRP: c_int = 0x5410;
+    pub static TIOCGPGRP: c_int = 0x540F;
 }
